Anpassung von Logging
